requirements.txt
kafka-python==2.0.2
requests==2.31.0
beautifulsoup4==4.12.2
pyspark==3.5.0
pymongo==4.6.1
elasticsearch==7.17.9
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
influxdb-client==1.38.0
google-api-python-client==2.92.0
